+++ libs/calendar/google/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib/google-calendar-connector';
+export * from './lib/google-api-error';
+export * from './lib/google-calendar-not-connected-error';